RAG_For_Certification/requirements.txt
faiss-cpu
sentence-transformers
PyMuPDF
google-generativeai
fastapi
uvicorn
python-dotenv
streamlit
sentence-transformers
faiss-cpu
groq
numpy